Disable MongoDB unit tests.
MongoDB will be completely removed as a primary backend. However, it
has become apparent that it may be useful to use MongoDB as a data
sink.

The idea is that all verification etc. takes place using efficient
internal data structures, but data is also exported write-only to an
external database such as a MongoDB cluster.